[feat] work version of generating chickens

+++ lab_3/src/utils/image_formatter.py
@@ -4,7 +4,7 @@ from pygame import Surface
 
 class ImageFormatter:
     @staticmethod
-    def scale_image(image: Surface, target_width: int):
+    def scale_image(image: Surface, target_width: float):
         original_width, original_height = image.get_size()
         scale_factor = target_width / original_width
         new_height = int(original_height * scale_factor)
